+++ xpreports/linux/report.go
@@ -25,7 +25,6 @@ type basereport struct {
 
 // BuildBase64bz2Report will return a compressed and encoded string of our report struct
 func BuildBase64bz2Report(conf *config.Config) (string, error) {
-
 	h, err := host.Info()
 	if err != nil {
 		return "", errors.Wrap(err, "Getting logged in users")
@@ -39,7 +38,7 @@ func BuildBase64bz2Report(conf *config.Config) (string, error) {
 		return "", errors.Wrap(err, "Getting root volume")
 	}
 
-	usernames, err := LoggedInUsers()
+	usernames, err := common.GetLoggedInUsers()
 	if err != nil {
 		return "", errors.Wrap(err, "Getting logged in users")
 	}
